Improve Email follow-ups configuration and fix missing logo (#125)
* Improve Email follow-ups configuration and fix missing logo

 - Fix some lists, the missing linebreak was causing Sphinx treating it as a paragraph.
  - Use a table to improve readability for global configuration options.
  - Add additional info about IMAP OAth (internal link)
  - Add missing GLPI logo. It's now shown on top of sidebar.
  - Minor fixes

* Change GLPI logo to white version

* Line blocks to prevent horizontal scrolling and br

This commit updates the email_notifications.rst file to, instead of the previously declared and used |br| html snippet, use the Sphinx 'Line block' feature.

Adding the correct line breaks on the line blocks prevent the horizontal scrolling on the table.

Because of this use, a margin was added below the last line. A custom CSS rule is added to prevent this behavior.

To add the custom CSS, the conf.py has been updated.

Minor grammatical fixes on other files.